Upgrade to Rails 3
If you're not bleeding you are too far away from the edge.

To get started all you should have to do is

gem install bundler # Manages dependencies
bundle install      # Get dependencies (inc. Rails 3), they are specified in the file Gemfile
rails server        # This is the new way to start the local server and so on